+++ truffle/src/com.oracle.truffle.dsl.processor/src/com/oracle/truffle/dsl/processor/bytecode/generator/AbstractBytecodeNodeElement.java
@@ -67,9 +67,11 @@ import javax.lang.model.util.ElementFilter;
 
 import com.oracle.truffle.dsl.processor.ProcessorContext;
 import com.oracle.truffle.dsl.processor.bytecode.model.BytecodeDSLModel;
+import com.oracle.truffle.dsl.processor.bytecode.model.BytecodeDSLModel.LoadIllegalLocalStrategy;
 import com.oracle.truffle.dsl.processor.bytecode.model.InstructionModel;
 import com.oracle.truffle.dsl.processor.bytecode.model.InstructionModel.ImmediateKind;
 import com.oracle.truffle.dsl.processor.bytecode.model.InstructionModel.InstructionImmediate;
+import com.oracle.truffle.dsl.processor.bytecode.model.InstructionModel.QuickeningKind;
 import com.oracle.truffle.dsl.processor.generator.GeneratorUtils;
 import com.oracle.truffle.dsl.processor.java.ElementUtils;
 import com.oracle.truffle.dsl.processor.java.model.CodeAnnotationMirror;
@@ -348,10 +350,28 @@ final class AbstractBytecodeNodeElement extends AbstractElement {
         ex.getModifiers().add(FINAL);
 
         CodeTreeBuilder b = ex.createBuilder();
+
+        if (parent.model.localAccessorsUsed.isEmpty()) {
+            BytecodeRootNodeElement.emitThrowIllegalStateException(ex, b, "method should not be reached");
+            return ex;
+        }
+
         buildVerifyFrameDescriptor(b, true);
 
         b.declaration(type(int.class), "frameIndex", "USER_LOCALS_START_INDEX + localOffset");
+
+        if (parent.model.loadIllegalLocalStrategy == LoadIllegalLocalStrategy.CUSTOM_EXCEPTION) {
+            b.startIf();
+            b.startCall("frame", "getTag").string("frameIndex").end();
+            b.string(" == ");
+            b.staticReference(parent.frameTagsElement.getIllegal());
+            b.end().startBlock();
+            BytecodeNodeElement.emitThrowIllegalLocalException(parent.model, b, null, CodeTreeBuilder.singleString("this"), CodeTreeBuilder.singleString("localIndex"), true);
+            b.end();
+        }
+
         b.startReturn().string("frame.getObject(frameIndex)").end();
+
         return ex;
     }
 
@@ -375,6 +395,12 @@ final class AbstractBytecodeNodeElement extends AbstractElement {
                         new String[]{"frame", "localOffset", "localIndex", "value"},
                         new TypeMirror[]{types.Frame, type(int.class), type(int.class), type(Object.class)});
         CodeTreeBuilder b = ex.createBuilder();
+
+        if (parent.model.localAccessorsUsed.isEmpty()) {
+            BytecodeRootNodeElement.emitThrowIllegalStateException(ex, b, "method should not be reached");
+            return ex;
+        }
+
         AbstractBytecodeNodeElement.buildVerifyFrameDescriptor(b, true);
 
         b.startStatement();
@@ -1023,7 +1049,7 @@ final class AbstractBytecodeNodeElement extends AbstractElement {
     }
 
     private static boolean isSameOrGenericQuickening(InstructionModel instr, InstructionModel expected) {
-        return instr == expected || instr.getQuickeningRoot() == expected && instr.specializedType == null;
+        return instr == expected || instr.getQuickeningRoot() == expected && instr.quickeningKind == QuickeningKind.GENERIC;
     }
 
     // calls dump, but catches any exceptions and falls back on an error string