+++ truffle/src/com.oracle.truffle.dsl.processor/src/com/oracle/truffle/dsl/processor/bytecode/parser/CustomOperationParser.java
@@ -853,6 +853,9 @@ public final class CustomOperationParser extends AbstractParser<CustomOperationM
             } else {
                 after.add(constantOperand);
             }
+            if (typeEqualsAny(type, types.LocalAccessor, types.LocalRangeAccessor, types.MaterializedLocalAccessor)) {
+                parent.localAccessorsUsed.add(type);
+            }
         }
         return new ConstantOperandsModel(before, after);
     }